intel: hwcomposer: Force FB Scaling for non-HDMI primary cases
Currently, I believe that there are only 2 devices uses this repo.
One uses HDMI_PRIMARY and is unaffected by this change (fugu) and
the other this fixes all kinds of problems (zenfone2).

In particular, it fixes the fact that hwcomposer allows a SPRITE
layer to be allocated in portrait mode which then cannot be used
after the display is rotated (which currently breaks chrome
when rotated).

If other devices start using this repo, we may need to revisit
this hard-coding of "true" and parameterize it or change the
higher level logic.

Change-Id: I44eca25430399f74f5e2f810d92ce71d2fd6f305